\name{getLattice.CellularAutomaton}
\alias{getLattice.CellularAutomaton}
\title{Lattice of Cells of a One-Dimensional Cellular Automaton}
\description{
This method extracts the matrix of cells from an instance of class \code{CellularAutomaton}.
}
\details{
\code{ca$getLattice()}
}
\value{
  \code{getLattice} returns a matrix of nonnegative integers. Each row of the matrix represents one
  generation in the evolution of the automaton.
}
\author{John Hughes}
\keyword{methods}
