\name{getNumberOfColors.CellularAutomaton}
\alias{getNumberOfColors.CellularAutomaton}
\title{Number of Colors of a One-Dimensional Cellular Automaton}
\description{
This method extracts the number of colors from an instance of class "\code{CellularAutomaton}".
}
\details{
\code{ca$getNumberOfColors()}
}
\value{
  \code{getNumberOfColors} returns an integer >= 2.
}
\author{John Hughes}
\keyword{methods}
