\name{getRadius}
\alias{getRadius}
\title{Radius of a One-Dimensional Cellular Automaton}
\description{
This method extracts the radius from an instance of class "\code{CellularAutomaton}".
}
\details{
\code{ca$getRadius()}
}
\value{
  \code{getRadius} returns an integer >= 1.
}
\author{John Hughes}
\keyword{methods}
