\name{getSteps}
\alias{getSteps}
\title{Number of Steps of a One-Dimensional Cellular Automaton}
\description{
This method extracts the number of steps (generations) from an instance of class "\code{CellularAutomaton}".
}
\details{
\code{ca$getSteps()}
}
\value{
  \code{getSteps} returns an integer >= 1.
}
\author{John Hughes}
\keyword{methods}
