\name{getTotalistic.CellularAutomaton}
\alias{getTotalistic.CellularAutomaton}
\title{Totalistic of a One-Dimensional Cellular Automaton}
\description{
This method extracts the setting of totalistic from an instance of class \code{CellularAutomaton}.
}
\details{
\code{ca$getTotalistic()}
}
\value{
  \code{getTotalistic} returns 0 or 1. A 0 indicates a general automaton. A 1 indicates a totalistic
  automaton. Outer-totalistic rules are not currently supported.
}
\author{John Hughes}
\keyword{methods}
